1、新增CSRF防护功能 2、升级tomcat到最新版本9.0.102 3、优化导出Excel日期格式双击离开后与设定的格式不一致问题 4、HttpUtils加入请求类型传参 5、优化空指针异常时日志无法记录错误信息问题 6、新增表格参数（自定义radio/checkbox的name值） 7、移除login_name更新字段 8、修复successCallback方法对于非特定表格类型无响应问题(IBG6ZD) 9、修复定时任务参数值带括号时异常问题(IBEBOX) 10、代码优化

+++ zhglxt-common/src/main/java/com/zhglxt/common/core/page/TableDataInfo.java
@@ -45,7 +45,7 @@ public class TableDataInfo implements Serializable {
      * @param list  列表数据
      * @param total 总记录数
      */
-    public TableDataInfo(List<T> list, int total) {
+    public TableDataInfo(List<T> list, long  total) {
         this.rows = list;
         this.total = total;
     }
